Move APFS list-stale-volumes, delete-volume code into library
Summary:
This is part of a stack that enables `eden du` to list and delete stale APFS
volumes on macOS.

In this diff, we're moving the dependencies of the `list-stale-volumes` and
`delete-volume` commands out of main.rs and into a new library crate that can
be used in edenfsctl.  This is a copy-paste job, but subsequent diffs will
add to this code and modify for testability.

Reviewed By: kmancini

Differential Revision: D39183490

fbshipit-source-id: a5e19a1ad900e925197ee1d9b31c8f9cb187c581