Add 2 types for scuba logging: local file and scuba table
Summary:
In tests we pass extra flag `--scuba-log-file` to redirect scuba logging to the file. However we do have `mononoke_derived_data` scuba table which is read from configerator config by number of services. This field is expecting the name of the scuba dataset, so we can't redirect this logging to the file atm. We have 3 options here:
 * Introduce another config field specifically for the scuba log file.
 * Introduce extra command line argument and pass it to each application which needs it for test
 * Mofify MononokeScubaSampleBuilder, so it can accept either scuba name or the file with `file://` prefix.
I went with the last approach as it seems cleaner and we do similar thing with configerator config paths.

Reviewed By: mzr

Differential Revision: D40015031

fbshipit-source-id: dfcd49270aa5ada618b9bfe61a3f6b277479c3dc