vfs: return NotFound for NotADirectory errors
Summary:
With workingcopy.ruststatus=True enabled the watchman tests were
failing when watchman reported a file path but the parent of the path had
actually been swapped with a directory since then. vfs.metadata() would throw a
NotADirectory error, which wouldn't be caught by the NotFound catch in status.

NotADirectory is actually an unstable Rust error and can't be matched on (see
link to the issue in the code). For now, let's have the vfs layer translate this
error into a NotFound error.

Reviewed By: quark-zju

Differential Revision: D40030949

fbshipit-source-id: 030dc0182ced7e1ed8404a7d7099a6e69b7369eb