backup commits to sapling-pr-archive branch
Summary:
Because `sl pr submit` relies on force-pushing to GitHub,
we run the risk of commits from previous versions of a PR
getting reaped, as GitHub periodically garbage collects
commits (and trees and blobs) that are no longer reachable
via a public branch.

To ensure we can always see previous versions of a PR in its
entirety, we merge the head of the development branch to a
special branch named `sapling-pr-archive-{username}`.
The idea is that this branch is not intended for development,
but to defend against GitHub's garbage collection process.

To support this, we add some new methods to `gh_submit.py`:

- `get_username()`
- `create_branch()`
- `merge_into_branch()`

These do what their names suggest via GitHub's GraphQL API.

Because `merge_into_branch()` has some "expected" error cases
(such as "branch does not exist"), we had to update the logic in
`make_request()` to ensure we can read the error back out of the
response to check whether it is one of these *expected* errors.

Over in `submit.py`, we introduce new functions to test the response
contents to see if it corresponds to an expected error:

- `is_already_merged_error()`
- `is_branch_does_not_exist_error()`

Finally, `add_pr_head_to_archives()` is where we put it all together
to merge the head of the development stack into the archive
branch, creating the branch, if necessary.

Reviewed By: quark-zju

Differential Revision: D39919020

fbshipit-source-id: 3822e7cb22d0b6f308a912fe98eec8106a87b5d9