identity: memoize sniffdir in Python
Summary:
Memoize sniffdir since this goes to the filesystem and could be called a lot in codepaths where the identity is not otherwise available (i.e. the "ui" object is not around).

I also aliased other pyidentity methods in the Python module so Python code can use the Python "identity" module consistently instead of the bindings.

Reviewed By: quark-zju

Differential Revision: D39867741

fbshipit-source-id: 85a8ebee78d9e8fab9c1777da963de8ce94fb0dc