repo: add ability to invalidate Rust repo requirements
Summary:
During repo init Python might change the repo requirements. We need
Rust to be aware of the new requirments. This happens during 'hg init --git'.

Ideally there'd be one source of truth for requirements, so we wouldn't need
such invalidation, but for now it's probably easier to invalidate. Later we can
remove python hg init entirely and stop needing this.

Reviewed By: bolinfest

Differential Revision: D40107731

fbshipit-source-id: 57a58bcb49e1ae8a6fc11d46ed85125a24ded2b9