regexmatcher: make RegexMatcher support '^'
Summary: Even though it's not necessary to use '^' in regular expressions (RegexMatcher checks for a match only at the beginning of the string by default), but we found many usages of '^' from the metrics. So we decide to support it instead of communicating with users and updating the current usages.

Reviewed By: quark-zju

Differential Revision: D40101477

fbshipit-source-id: 5a13804f65a15b4d923bb6d8ddd1e78669569ab7