Preserve root by default on mutable-renames add cmd
Summary:
I made a script to call this code, but by accident it called it with an empty string for src and dst path, which made it try to mark all files in the repo as moved.

It didn't cause any problems, but it could have, so let's make it hard to make that mistake.

Reviewed By: markbt

Differential Revision: D40065862

fbshipit-source-id: 8d930f7af868552a85b9fffd28a097f6b4c7756d