serve reads from OverlayBuffer work queue
Summary:
To increase performance, we need to allow for reading from the write queue directly. This is implemented by mapping an inode number to an `Operation` (a pointer to the `Work` payload and operation type (read or write)). There are two "states" that `Operation` can have:
* waiting
    * The underlying `Work` object is waiting to be processed by the worker thread
    * The underlying `Work` object is on the `work` vector inside of class member variable `state_`
* inflight -  The underlying `Work` object is currently processed by the worker thread
     * The underlying `Work` object is currently being processed by the worker thread
     * The underlying `Work` object is on the `work` vector inside of the `process()` method

A single `Work` object can only be in one of the two `work` vectors, meaning that an `Operation` can only be in one of the two `operation` maps.

An `Operation` always has a corresponding `Work` object, but a `Work` object might not always have an `Operation` referencing it. This is if there are multiple `Work` objects for the same inode number. There will only be an `Operation` for the most recently observed `Work` object.

Here is a visual:

{F772371636}

Reviewed By: chadaustin

Differential Revision: D38761415

fbshipit-source-id: 089c90379a2a784ae37700b3f0ce64bd05596169