bail early when we hit an unmount timeout/exception
Summary:
Removing state underneath a running EdenFS mount is a bad idea. In the past, we rarely hit the timeout so this was fine. We've now seen more and more people hit this timeout (due to thrift queue timeouts).

Let's avoid messing up the EdenFS state even further and instead give users methods for remediating a hanging mount.

Reviewed By: fanzeyi

Differential Revision: D40113414

fbshipit-source-id: 35f2c11717be346d2c14b3392a7ee6d98b603483