Preserve case of glob matches
Summary:
On case-insensitive but case-preserving file systems, we currently clobber the case of a glob match and synthesise a path from the *pattern*'s case - e.g. if the pattern is `foo` and the file `Foo` exists, we incorrectly return `foo` as the match. This erases information that may be important to clients.

Here we change Eden to consistently preserve the case of the path as stored in the FS.

Reviewed By: chadaustin

Differential Revision: D39892007

fbshipit-source-id: 9d857ebdee917d0045ab800ef7094b85bf129722