include RuntimeError in except clause
Summary:
D40132731 suggested I remove a naked `except:`,
so I took my best guess and changed it to
`except subprocess.CalledProcessError`, but apparently
`shell.py` looks at the resulting proc and if it has a non-zero
exit code, it throws a `RuntimeError`, so we need to
catch that, as well.

Incidentally, `git config --get` appears to return a non-zero
exit code if the config name does not exist, so this is
expected here.

Reviewed By: evangrayk

Differential Revision: D40165764

fbshipit-source-id: dc0bfd66ae42d23a58f24880670e847f589cbc1c