execute cloud sync in foreground after `hg cloud hide` if it
Summary:
execute cloud sync in foreground after `hg cloud hide` if it runs
for the current workspace. If it runs in background, this may be a bit
confusing not to see actual hiding if the command was executed for some commits
that are present locally. Also, the background sync maybe disabled, in this
case the effect from the hiding will not be seen.

If hiding was executed for another workspace rather than for the current one, the background cloud sync call is not needed. Let's disable it.

Reviewed By: markbt

Differential Revision: D40223525

fbshipit-source-id: aeedc1cb85ba8236f4649b67520ff119f89e06d5