do not skip the '--full' option in `hg cloud sync` help
Summary:
do not skip the '--full' option in `hg cloud sync` help

The option should now works for all users (thanks to the "old commits cleanup"
project) and it is not horribly slow nowadays due to unhydrated bundles and
other optimizations.

So, we shouldn't really skip it from the help, because it is a way how users can learn about this option.

Differential Revision: D40259500

fbshipit-source-id: dd254e0ceea88648633bcefe0297d38e50b4a43a