Remove shared references for configuration
Summary: The current API of `MononokeApp` exposes `RepoConfig`, `StorageConfig` and `CommonConfig` as shared references rooted under the single `MononokeApp` instance. However, post config hot-reloading the configurations would be dynamic and can change at runtime. If callers maintain a shared reference to the configurations, this won't be possible. This diff removes the shared reference and replaces it with an `Arc` reference so even if the underlying config gets refreshed the callers will have a pointer to an outdated config which will be freed once the reference goes away.

Differential Revision: D40073990

fbshipit-source-id: a8d0210acd6a2d3dff65ba06cbc1b08d0219cbd9