check the result of the update operation in `hg cloud hide` and `hg
Summary:
check the result of the update operation in `hg cloud hide` and `hg
cloud move`

The result of the operation wasn't checked in previous existed implemetation of
the `hg hide` command and it wasn't aslo checked in the implementation of `hg
move` command

The updates don't need to be transactional. It's ok if the second one fails and the first one succeeds. In this case it will be rather a 'copy' operation rather than a 'move', but it still can be fixed by a retry (as we advice here in the diff) due to the deduplication logic we have in place when we update the destination workspace. So, the changes won't be applied twice, and it will not fail if we try to apply twice, we will just ignore the duplicated update.

Reviewed By: markbt

Differential Revision: D40295995

fbshipit-source-id: 5e612bd97f5601e4ec2f5ea8e7316d8c115d1615