Make repo_client/unbundle not depend on mononoke_api
Summary:
I want to add a dependency from `mononoke_api` to `unbundle` on a later diff, but that would currently generate a cycle.

The only reason `unbundle` depends on `mononoke_api` is to use its `Repo` container. But it could instead just use facets to avoid this issue (and help with the BlobRepo effort as well).

This diff makes that code a little more facet-y, and removes the dependency.

Reviewed By: markbt

Differential Revision: D40307246

fbshipit-source-id: 6c9ea9f2b3195b81453eb91a099692b64fc9ac86