immediatefuture: make ready when the SemiFuture is ready
Summary:
In the case where the promise associated with a SemiFuture is set after the
ImmediateFuture construction, EdenFS will still attempt to push the SemiFuture
to an executor. Since this is quite expensive, let's handle this case by
considering the ImmediateFuture as ready.

In very high fanout computation (like glob), this is expected to reduce the
amount of futures pushed to the Thrift executor. This is expected to help
reduce the amount of Queue Timeout in `glob(**/*)` cases.

Reviewed By: genevievehelsel

Differential Revision: D40312497

fbshipit-source-id: e48efac2e4f2640b6a7ad50957ee7c723772ec81