identity: combine user and repo identities when sniffing
Summary:
Now when sniffing the repo identity, the returned Identity combines the sniffed repo identity with the user identity attributes of the default identity. This means when using "sl" in a ".hg" repo, calling identity::sniff_dir() will give you a mixed identity with dot_dir of ".hg" and cli_name of "sl".

This is mainly to keep things consistent so all identity objects you obtain in the code can be used to format user facing messages without having to think about where the identity came from.

Reviewed By: quark-zju

Differential Revision: D40279955

fbshipit-source-id: 1874a608c1c9ab9f13a5936f8c6b5b93fd1683bf