identity: separate into user and repo identity
Summary:
Some attributes of the identity are strictly tied to the repo (e.g. ".hg" vs ".sl" directory), and some are user facing and should be tied to whether the user invoked "hg" or "sl". Separate the identity fields into separate structs to make the delineation clear.

I'll use this separate in the following diffs to clean things up.

Reviewed By: quark-zju

Differential Revision: D40279962

fbshipit-source-id: 0fbce016f92b852ed5489b9bf770df161f3c90e0