remove usages of `untrusted` option for config related code in Python
Summary:
Across the Python part of our codebase we seemed to have previously supported the `untrusted` parameter in various config related methods. Since these parameters were not actually used, they can be deleted.

This in particular affects the `config` command, which previously seemed to do something with the `--untrusted` option.

Reviewed By: quark-zju

Differential Revision: D40355828

fbshipit-source-id: f467e564f59564fb63ef4a194957a79971303656