workingcopy: store VFS instead of root
Summary:
The VFS object represents the working copy filesystem. It stores useful
information like whether the file system is case sensitive. In a later diff
we'll need this information when creating the treestate and matchers, so let's
store the VFS on working copy and filesystem objects instead of the root
PathBuf.

Reviewed By: jordanwebster

Differential Revision: D40164605

fbshipit-source-id: c0e4c74525b5b30826fb70ce9e177476a750f58b