Run case sensitivity integration tests exhaustively on each platform
Summary:
Refactors `glob_test` and `casing_test` to run with every possible case sensitivity setting (sensitive, insensitive, and OS-determined) on every platform where we run the tests. More generally, any integration test can now be decorated with  `testcase.eden_repo_test(case_sensitivity_dependent=True)` to get the same behaviour.

This makes it easier to iterate on case-(in)sensitivity features while developing Eden on a single machine/OS (that may or may not be case-sensitive itself), and removes the hacky OS sniffing that we were previously doing for `glob_test` and `casing_test` in favour of always asking Eden what the real case-sensitivity setting is. This is powered by the CLI changes in D40180643.

Reviewed By: xavierd

Differential Revision: D40180642

fbshipit-source-id: e3dadd274e3b82dc7e6ebf465ae70287ea623594