Allow setting and querying case-sensitivity
Summary:
Currently, the case-sensitivity of an Eden checkout is set at clone time based on the OS: Linux = case-sensitive, macOS and Windows = case-insensitive. This works fine but leads to patchy integration test coverage of case-sensitivity features, since coverage is tied to the OS the tests happen to be running on. Relatedly, there's no clean way to query the case-sensitivity of an existing checkout other than reaching into `config.toml`.

Here we introduce a pair of hidden `eden clone` flags to allow changing case sensitivity, and add a `case_sensitive` boolean field to the output of `eden info`. Because we don't want end-users to rely on this, setting case sensitivity to anything other than the default on each platform triggers a warning.

I will use this in a separate diff to refactor certain existing integration tests (namely `casing_test` and parts of `glob_test`) to run with all possible case-sensitivity settings on every platform.

Reviewed By: xavierd

Differential Revision: D40180643

fbshipit-source-id: 4b4c935f3c9bb40f8e3564383a91dc38ea25eab5