wal multiplexed: add perf counters
Summary:
This enables ODS logging for the number of blob writes and reads, its size and multiplexed store operation latency.

That's basically the same logic as in old multiplex, however we could improve it later by revising the meaning of the counters, for example, `BlobGetsMaxLatency` currently logs the overall latency, although we may want to log the actual max in between of all the inner gets.

Reviewed By: yancouto

Differential Revision: D37613228

fbshipit-source-id: 231ce6c8c2326e6f9fc51c0df400033e6acd7ffc