inodes: better handle concurrent file to directory during checkout
Summary:
When a file is replaced by a directory while checkout is ongoing, EdenFS would
add an error conflict which is then displayed on the CLI to the user. By making
the conflict a non-error, Mercurial has a chance to remediate the conflict a
second time which should succeed or fail with a more meaningful error.

Reviewed By: chadaustin

Differential Revision: D39699126

fbshipit-source-id: e105ab26ecc7a25dbfd97bfd3cb34c74364437d4