bonsai_*_mapping: remove auto_impl
Summary:
These mappings are nearly always stored as `Arc<dyn Mapping>`, so we can remove
the generics and switch to always expecting dynamic dispatch.

The generic versions of `MemWrites` and `Caching` mappings can be converted to
always expecting `Arc<dyn Mapping>` as the inner mapping (in practice they
always got this except in some unit tests).

The cache implementations no longer need to be generic over the inner cache type.

We can remove `Clone` from the mappings themselves to ensure we only ever clone
the `Arc`.

Differential Revision: D40357649

fbshipit-source-id: 988f96d9b68e827a4afedfeb785c9731783cc6b6