stop using BUCKVERSION environment variable
Summary:
When Eden kills Buck programmatically (for redirect subcommands and eden rm), we set a number of environment variables to ensure Buck functions correctly. One of the environment variables we set is BUCKVERSION. We set this env var to "last" in order to avoid downloading new versions of Buck when we're simply trying to kill buck in the first place.

This environment variable has started causing issues for stale repos. A stale repo is any repository that's checked out to a commit that's sufficiently old (we currently don't know what age this is, but typically 4-5 day-old commits seems to trigger the issue).

We should stop using this environment variable until we figure out a workaround or a fix for Buck

Reviewed By: mshroyer

Differential Revision: D40517904

fbshipit-source-id: 419d09814dc315ec79410cafebbb50b172992fd9