watchman: don't edit the dirstate every Rust status
Summary:
The Rust watchman code was flushing the dirstate after every status.
This is excessive and causes things like VS Code to trigger, which then run hg
status, and forms an infinite loop.

To fix it, we A) only flush if the treestate is dirty, and B) don't mark things
as needs_check which are already needs_check.

Reviewed By: muirdm

Differential Revision: D40532717

fbshipit-source-id: 95b07c1b8feb86af4c7556394992d255bfb553ff