Add (fake) retryable SQL queries
Summary:
**Context**: We want to add retries to all our SQL queries automatically, when the error is retryable.
To do this, let's add a `queries_with_retry!` macro that works the same as the `queries!` macro, but does retries automatically when possible. I'll do that over this stack.

This diff creates the basic macro, which for now works only for `read` queries, and doesn't actually do any retries, just does the query normally. I'll make the macro more powerful over the next diffs.

Reviewed By: mitrandir77

Differential Revision: D40105656

fbshipit-source-id: 344e99cd7c6075fe38d91f0728d8e84400e98475