Use hot-reloading configs in MononokeApp
Summary:
Due to the work done in the last diff, we can now directly use `MononokeConfigs` for getting the most up to date configuration in `MononokeApp`. The methods `repo_configs()` and `storage_configs()` on `MononokeApp` will now return the most recently updated version of `RepoConfigs` and `StorageConfigs`. The return type is an `Arc` so that even if the config is updated the user will end up holding a safe but outdated handle.

Fetching the repo and storage configuration is also supported at the repo level (i.e. using the `Repo` facet). The difference between the two fetches is that `MononokeApp` based fetches will always provide the latest configuration value while the `Repo` facet based fetches will provide the configuration that was used at the time of repo construction.

Note that this diff only implements hot-reloadable configs. Updating the repos based on the reloaded configs will be done later in the stack.

Reviewed By: markbt

Differential Revision: D40298571

fbshipit-source-id: 7036a7f45ea9da34f214d57f823876900bf8985e