Implementing repo-reloading in MononokeApp on config change
Summary:
This diff implements the repo-reloading part which needs to be triggered once `MononokeConfigs` detect a configuration change. The amount of code change is small but there are certain repercursions of the current design.

Before reloading, we check which repos are currently being served by this server because the actual list of repos can vary due to sharded setting. Then, we first construct the repos with the updated config before putting them in place of the outdated repos. This means that for a short while during the construction we would have twice the original number of repos. As a result, our servers should be capable of supporting 2X the number of repo that it is configured with.

It is also possible (though highly unlikely) that during the time the `MononokeConfigUpdateReceiver` reloads the repos, the `ShardedProcessManager` could add or remove repos. This would result in adding dropped repos or removing newly allocated repos from the server. This would then be detected by ShardManager consistency check and fixed by reinvoking the `addShard` or `dropShard` callback.

Reviewed By: markbt

Differential Revision: D40412930

fbshipit-source-id: b8755ad403204e81ec5b55dc5dbf55d1754f3c78