Replace MononokeRepos with Arc<MononokeRepos>
Summary: `MononokeRepos` is meant to function as a shareable structure that gives access to dynamically reloadable repos. As such, no caller is meant to own `MononokeRepos` but instead are expected to have an `Arc` handle to it. This diff updates the `open_mononoke_repos` method in `MononokeApp` to achieve that objective. This is also important for hot-reloable configs which would require an `Arc` handle to `MononokeRepos`

Reviewed By: markbt

Differential Revision: D40392067

fbshipit-source-id: 92c5d500417882218e7c0868d6b0b6eabb22af34