scs_server: fix clean shutdown
Summary:
Clean shutdown for the SCS server seems to have been broken for a while.

Since the thrift service handler runs on the folly executor (triggered by starting service framework), we don't need a service handler running on the tokio executor.  However, we still want to be able to use the shutdown handling code from `serve_forever`, so we need a dummy service future to pass to that.  Previously we passed the `monitoring_forever` future, however since this never exits, we will always wait until the timeout when shutting down the service.

Instead, spawn the monitoring future separately, and create a real dummy future
that waits until we are signalled to shutdown before completing.

Reviewed By: yancouto

Differential Revision: D40587920

fbshipit-source-id: 270a1da8a95e1ea83115cb005c53a52b41e60b09