identity: prefer ".sl" over ".hg"
Summary:
For various reasons, sometimes a ".hg" directory gets created in a ".sl" repo. Previously that screwed everything up because we sniffed for ".hg" first and ignored the ".sl" dir from then on.

With this change, we sniff for ".sl" first, so we have some protection as we chase down ".hg" vivifications.

Reviewed By: bolinfest

Differential Revision: D40498410

fbshipit-source-id: 28a92bec71f56294c9109b4cce052d4d32b361b1