service: implement a basic invalidateKernelInodeCache on Windows
Summary:
This is far from being complete, and far from being perfect. It merely calls
PrjDeleteFile without looking at the inode state, or the inode type, and
doesn't attempt to reset the fsRefcount.

This is added as it proved to be useful to repro the ProjectedFS bug by
completely excluding any EdenFS/Mercurial logic.

Reviewed By: genevievehelsel

Differential Revision: D40574897

fbshipit-source-id: 749495cb8f55f81f9203ae7fa9726cb9b734d0f5