Create push redirector on repo_land_stack
Summary:
**Context**: See D40384759

In order to do pushredirection from SCS, we have to extract and use it from wireproto handler, without reusing that whole thing because we plan for it to go away.

This diff creates the `PushRedirector` object from the config, to conform with the signature change from the previous diff, but still doesn't use it. That is done on the next diff.

Differential Revision: D40384758

fbshipit-source-id: 6e83ce764b44d8fe4e84ae0c70259da477470c3e