Implement ApfsUtil struct for managing eden APFS volumes
Summary:
This is part of a stack that enables `eden du` to list and delete stale APFS
volumes on macOS.

This diff refactors some existing Rust APFS volume management code into a new
struct, which has a DiskUtilCommand provided as a dependency in order to make
its behavior testable.  (A later diff will add tests when modifying ApfsUtil so
that unmounted APFS volumes with hashed names are not incorrectly counted as
"stale").

Reviewed By: fanzeyi

Differential Revision: D39283675

fbshipit-source-id: 158b1b913e22ed17b172783c43e2d2171e7c953e