wal multiplex: define new config
Summary:
This diff declares a new config for the WAL multiplexed blobstore. It is different from the current multiplex in the way it defines write/read quorums.

The main reason for a separate config is ability to iteratively switch repos and different parts of the Mononoke tools from the old multiplex to the new one.

Reviewed By: yancouto

Differential Revision: D40108825

fbshipit-source-id: 9dae447557702474b6427ad247164a879da90861