Prevent APFS volumes incorrectly being counted "stale"
Summary:
This is part of a stack that enables `eden du` to list and delete stale APFS
volumes on macOS.

The `eden_apfs_mount_helper list-stale-volumes` command can currently
incorrectly list volumes as "stale" if they have a hashed name (corresponding
to a redirection mount point with a path longer than 120 characters) and if
they're presently unmounted.  This happens because in order to determine
whether a volume is stale, the preexisting `list_stale_volumes` implementation
checks whether the volume's name contains the path of any known eden checkout.

This diff provides a new `list_stale_volumes` implementation that takes a set
of configured redirection mount points instead of a set of checkouts.  From
these it computes the set of all possible non-stale APFS redirection volume
names, hashing if necessary, in order to exclude these from the list of stale
volumes.

Because `eden_apfs_mount_helper` doesn't have logic to read the set of
configured redirections from disk, for now it still uses the old implementation
which is renamed to `list_stale_volume_unsafe`.  Going forward, once `ApfsUtil`
is landed we can make the Rust-based `redirect` command use the new
implementation directly instead of shelling out to `eden_apfs_mount_helper`.

Reviewed By: fanzeyi

Differential Revision: D40329861

fbshipit-source-id: 130468e24e962a9c083e04d5bc9174d83b03bf46