git clone: set up repo config earlier
Summary:
Now we set the repo config (in particular paths.default) as we initialize the repository. The ever important remotefilelog.reponame is inferred from paths.default, so everything is easier if paths.default is there from the get-go.

For example, updating the repo's config later left the repo._rsrepo object with a stale config with the wrong reponame.

Reviewed By: bolinfest

Differential Revision: D40474871

fbshipit-source-id: 6973b1c7c7041df96929647b11e803eb53a31ecd