cli_rs: remove skipserializing_if and add merge_option
Summary:
This diff fixes the issue that `eden config` does not show `[core]` section.

The underlying cause is that, when merging values, stack-config does not know how to merge Option values. By default it always take the new value if it is provided.

This is an issue since we are also using `stack(default)` with `edenDirectory`. So it is always going to have a value, None, which will be considered a value in the upper layers.

Reviewed By: kmancini

Differential Revision: D40605177

fbshipit-source-id: fae04b108eec2f480476c0b5a77e20c7e825ba91