subscribe: add a timer to reduce likelihood of event miss
Summary:
During my testing, it looks like there will be cases when somehow we missed an event from the command line, and no event will be delivered because no one is calling `getCurrentJournalPosition` (in reality this is probably very unlikely since Watchman will be calling it all the time).

But better to be careful than stuck in the state where no changes are being reported. This diff adds a timer (defaulted to 15 seconds) that will triggers a `getCurrentJournalPosition` call to EdenFS. This will be safe since we will not be generating a changed event if the `sequenceNumber` has not changed.

Also, better name is welcome.

Reviewed By: MichaelCuevas

Differential Revision: D40453342

fbshipit-source-id: 60e2aae5ba1f8f79d2b5c11b2c42eaec5e25dee3