subscribe: do not report .hg changes
Summary:
The current subscribe command will be reporting events if anything in `.hg` directory changes. This is by design but it unfortunately does not work well with our use cases as it could trigger an infinite loop.

This diff teaches the subscribe command to use `getChangesSince` method to examine the changes happened during the time frame, and decide if it wants to report an event based on what has changed.

Reviewed By: MichaelCuevas

Differential Revision: D40190204

fbshipit-source-id: 9173dbc02efbf544801e7276299247319334533c