Test timing of importing a large merge commit
Summary:
TL;DR:
* Generate a repo with 100 commits-long branch that gets merged into master and import it
with `remote_gitimport`.
* Repeat with 1000 commits-long branch.
* Show that the amount of work done in each thrift call to SCS doesn't increase linearly
  with the number of commits in the branch being merged.
* Show it would without D39810825
* We add two tests because the test with 100 commits can be used for quicker iteration and
  allows us to display the entire relevant outputs.
  The test with 1000 commits allows us to truly test the scale where even if the speed of
  deriving data was optimized, it should still catch any regression.

Background:
Recently, a new endpoint was added to SCS to derive data for a batch of commits (D39810812 (https://github.com/facebookexperimental/eden/commit/4e8d2b308772976c3e5c63a08004281a9c782e67))
That endpoint was then used to derive fsnodes data on batches of 20 commits at a time
while importing git repos with remote_gitimport (D39810825)

These changes were made to solve a problem described by Simon Farnsworth in
[this post](https://fb.workplace.com/groups/scm.mononoke/posts/1771878673174700/)

At the time of importing a merge commit - or more precisely, at the time where mononoke
creates the bonsai changeset for this merge commit, to evaluate whether there are any conflicts
between the branches being merged, SCS needs to derive the `fsnodes` data for the merge commit.

Since data can only be derived for commits whose ancestor already had the same data derived,
if the `fsnodes` data for all ancestors of the merge commit must be derived at that point.

This generates an amount of work that's O(number of changes in branch).
Since this is potentially unbounded, we will run into cases where the 460s timeout in production
won't be enough to derive all of the necessary data.

With this change, we want to ensure that the amount of work being done when creating the changeset for
the merge commit is bounded.

The `REMOTE_GITIMPORT_RECV_TIMEOUT` exposed in D40512225 represents the time the `remote_gitimport`
client will wait for an answer from the SCS server before the thrift connection to it times out.

Deriving the fsnodes data for 1 commit takes a few ms depending on circumstances, so doing
the work for 100 commits should always take more than 3s and trigger a timeout.
Deriving data for 20 commits should take around 20s, so deriving a batch of 20 commits should not
exceed the timeout of 30s.

This test shows that deriving the `fsnodes` data as is done in prod at the moment is enough to avoid
the potentially unbounded amount of work at `create_changeset` time that was facing us before.

Additionally, by showing the SCS server log output, we can see that `fsnodes` data is derived in
batches every 20 commits until the merge commit, where anything that still hasn't been derived
gets derived.

Note that `skeleton_manifests` and `changeset_info` data gets derived more often. That is because
unlike `fsnodes` data, that data is needed for the parent commit at the time a child commit is created
by SCS whether this commit is a merge commit or not. That means that we never have the opportunity
to build up a backlog of missing derived data for these types and batch deriving wouldn't be applicable
in this situation.

Reviewed By: mitrandir77

Differential Revision: D40512224

fbshipit-source-id: 8e304df4ba9b5abe87eaa430f07501be49f0b34e