populate the Yarn offline-mirror
Summary:
Recall that `sapling-cli-ubuntu-20.04.Dockerfile` and
`sapling-cli-ubuntu-22.04.Dockerfile` are used as part of the
respective `sapling-cli-ubuntu-xx.yy-image.yml` GitHub actions
that are used to create Docker images for the other actions so
that tools and dependencies do not have to be installed from
scratch each time.

Note that this existing line in the `Dockerfile`:

```
RUN yarn config set yarn-offline-mirror "$HOME/npm-packages-offline-cache"
```

was meant to set up an offline cache so `yarn install` in subsequent
builds would be faster. Unfortunately, it appears we were not
actually populating the offline cache because we were not
specifying the `--prefer-offline` flag to `yarn install`, which this diff fixes.

Previously, we were also defensive in the call to `yarn install`:

```
RUN [ -f /tmp/repo/addons ] && yarn --cwd /tmp/repo/addons install || true
```

But now the build will fail if `install` does not succeed, as the previous
approach made it too easy for problems to go unnoticed.

I updated the `.yml` files by running:

```
~/fbsource/fbcode/eden/oss/ci$ buck2 run :gen_workflows -- ../.github/workflows/
```

Reviewed By: evangrayk

Differential Revision: D40658823

fbshipit-source-id: 2813641b746654ac06e15ac151f1c017d8e39622